requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose==3.3.0
pydantic==2.10.3
pydantic[email]==2.10.3
python-multipart==0.0.20

# testing
pytest==8.3.4
pytest-cov==6.0.0
httpx==0.28.1
